fix: CI 修复 - clippy 警告 + 测试断言
- 移除 opc_learning_hook.rs 中冗余的 as f64 类型转换
- 更新 opc_workflows 测试断言 count 75 -> 76
- rustfmt 格式化执行器和命令模块

+++ src-tauri/src/commands/opc_learning_hook.rs
@@ -154,7 +154,7 @@ pub async fn try_auto_learn_workflow(
     };
 
     let quality_score = compute_quality_score(result);
-    let quality_score_100 = (quality_score * 100.0) as f64;
+    let quality_score_100 = quality_score * 100.0;
 
     info!(
         "[opc-auto-learn] 触发自动学习: industry={}, template={}, quality={:.1}",
